Also reset the compiler connections within the compiler classes
